track_script - kill the process group, not just the process.
It was possible for a script (specifically burst_buffer.lua) to spawn
processes, but then when the script was killed track_script only killed
the parent process in the process group. This resulted in the other
processes continuing to run. If this happened on slurmctld shutdown, then
slurmctld would hang forever due to track_script doing pthread_cancel()
on the slurmscriptd thread that was tracking the script, and then that
thread would never be able to send a response to slurmctld once the
processes were finally complete.

This is all solved by making track_script kill the process group, not
just the process.

Bug 12429