Stop clearing the CRON_JOB flag when loading job state.
Commit 507508257f fixed a bug where older jobs being loaded in Slurm
20.11 were marked as CRON jobs if those jobs had invalid
dependencies. This is because CRON_JOB reused the now-unused
INVALID_DEPEND flag spot in 20.11. The solution was to clear the
CRON_JOB bit flag spot in all 19.05 and 20.02 jobs loaded from state,
starting in 20.11.

However, starting in 21.08, 20.11 and 21.08 jobs were now erroneously
clearing the CRON_JOB flag unconditionally on job state load due to a
bad 20.11-to-master (21.08) merge in commit cbede5e65f.

Fortunately, the current 20.11 works as commit 507508257f intended and
was not affected by this error.

Bug 12451