Fix slurmctld min_nodes estimation for certain config and requests.
The max number of cores on a node in a partition is incorrectly calculated
in build_part_bitmap() when the partition contains multi-socket nodes.
Since it doesn't account for the number of sockets the resulting max number
of cores is lower than the actual.

This part_record_t max_core_cnt is used in _pack_default_job_details() when
min_nodes is estimated based upon num_tasks and ntasks_per_core. This in
turn results in incorrect node count estimations for certain jobs to be
displayed client-side.

Bug 12212