Simplify accounting query after a7e98c4277.
No functional change.

The aforementioned commit makes sure that no jobs with het_job_id=0 are
returned from the inner select statement (so only HetJobs are returned).
This makes checking that het_job_offset != NO_VAL (as an original way to
ensure only HetJobs are filtered) redundant and irrelevant. This in turn
makes a pair of parentheses unneeded, and thus are also removed.

Bug 11516