Preserve job start info in the database for a requeued job that hadn't registered the first time in the database yet.
If a job ran and got requeued before its start got sent to the database
(the job's db_index is zero), then all of the information in the "start"
database record got lost. Two examples of information that got lost are
the job's start time and the job's name. By losing the start time we
also lose elapsed time, which affects the usage tables. Without the job
name, the job name was overwritten to be "allocated".

We fix this by sending an update to the database for the requeued job
before we reset the job's db_index to zero. But at this point the job is
pending, and for pending jobs we send a zero as the start time to the
database. So we also have to tell jobacct_storage_g_job_start() that the
job was requeued and we should preserve its start time instead of sending
zero as the start time.

Bug 8722