fatal() if TaskAffinity=yes in cgroup.conf.
The cgroup.conf parameter TaskAffinity=yes is broken in several ways in
the 21.08 release. This is due in large part to the overhaul of the cgroup
infrastructure. The affinity offered by the TaskPlugin task/affinity (which
can stack alongside task/cgroup) has been recommended for some time.

Rather than try to rework the TaskAffinity=yes infrastructure in 21.08,
fatal() when configured instead and inform the site of the issue. SchedMD
was already preparing to remove this in 22.05, but further problems that
have been discovered leads us to remove support for this option now.

Bug 12564.

Signed-off-by: Tim Wickberg <tim@schedmd.com>