Handle job resize better in the database.
A few of things are happening in this commit, but they need to be
changed at the same time.

1. Have the slurmctld handle the end of a resized job instead of having
the slurmdbd do double duty on starting the new sized job and end the
old one.  Ending the old one messed up accounting on things.
2. When sending a new job start on a resized job don't do a temporary
submit_time.  This isn't needed anymore since the slurmctld is sending
the job_comp message.  Now resize_time is the only thing used in the
database and submit_time is ignored.
3. I also know that job_start_direct can't be used here,
I am setting the db_index = NO_VAL64. Removing that comment.

Bug 12143