Exclude currently running, resized jobs from the runaway jobs list.
If a running job was ever resized, then it would erroneously show up
as a runaway job in `sacctmgr show runaway`.

When a job is resized, the database will create a new record with the
job's submit time set to the resize time. However, the job's submit time
in the ctld will show the original time. So when checking for
runaway jobs we wouldn't find the resized job in the ctld since
it's submit times wouldn't match up with the submit times in the
database. So simply check the db's job submit time against the ctld's
job submit time OR resize time to properly filter out resized records
from the runaway jobs list.

Bug 6948