Fix checking if the dependency is the same job for array jobs
When adding a dependency to a job, we check if the dependency is the same
job as the job it is being added to. For example, the following is invalid:

scontrol update jobid=15 dependency=afterok:15

Array jobs use a different syntax, such as 16_* (or just 16) for the whole
array job or 16_1 for a single task in the array. In either case, we weren't
checking if the dependency was the same job correctly. For example:

scontrol update jobid=16_1 dependency=afterok:16_1

This should fail, but previous to this commit this didn't fail because the
actual job id of job 16_1 was not 16 (unless 16_1 was the meta record for
the array job, in which case its job_id matches the array_job_id).

This commit uses the job_id and array_task_id of the dependency string
(16 and 1, respectively, in this example) and whether or not the job
being updated is a specific task of an array job to fix this.

Bug 8492