Set array_task_id when the last job record is split from the job array
Make sure that the meta record for a job array gets array_task_id set
correctly after splitting off the last record in the array. This fixes
issues with updating job dependencies for that job within the job array.

If we don't do this, then array_task_id would stay at NO_VAL for the
last job record of the array until the job got scheduled.

This is a continuation of the previous commit.

Bug 8492