Testsuite - update test 39.18 subtest 2.6 to reflect current behavior.
Commit 3bd0778f9a improved the existing behavior to make more sense.
So now change the expected output to match this.

It makes sense that all three GPUs should have only the HAS_FILE and
HAS_TYPE flags ($flags_file_type), because those are the default
auto-detected system flags in test 39.18, gres.conf did not specify any flag
overrides, and system flags trump default gres.conf flags. However, before
commit 3bd0778f9a, the final GPU dev4 also contained ENV_NVML, ENV_RSMI,
ENV_OPENCL, and ENV_DEFAULT ($flags_default_type), which means it was
preserving the default gres.conf flags for dev4 and was not properly
deferring to the system GPU's flags for dev4. This was a low-priority,
edge-case bug. After commit 3bd0778f9a, this unexpected result was fixed,
due to improvements in the way the system and conf GRES are matched. So now
we are fixing the test.

Bug 12126