gres/gpu - fix configured/system-detected GRES match for some cases.
Before, we were sorting GPUs by File in descending order right before doing
conf-to-system matching. This was so that config GPUs with a NULL File would
have the lowest matching priority.

However, after commit 3bd0778f9a, this caused issues because
_normalize_sys_gres_types() was no longer clearing out the type, causing all
GPUs to be available during matching and making the intermediate descending
order relevant to the selection. This was a problem when there were more
system GPUs than config GPUs. So when we sorted GPUs in descending GPU File
order before matching, we would end up selecting GPUs from the opposite end
of the list than expected, and leave the first GPUs as unused instead of the
last GPUs.

We want to sort GPUs according to File in ascending order, but we also want
NULL File GPUs to be last. So that is what this commit does.

This fixes test 39.18 subtests 0.[38,39], which were failing ever since
commit 3bd0778f9a.

Bug 12126