Allow jobs to pack on nodes already rebooting with the desired features
With a node_features plugin active, there was a race that could cause a
job to pend forever due to a lost reboot request

* One job requests a node feature.
* reboot_job_nodes() queues a reboot request.
* The slurmd gets that reboot request. The node reboots but slurmd
  hasn't started yet.
* Another job requests the same feature and is scheduled on the same
  node as the first job. reboot_job_nodes() sends another reboot request
  to that node because the node's active features haven't been updated
  yet. But, because the slurmd hasn't started yet, the reboot request is
  lost.
* The slurmd comes back online, but the reboot time is *before* the
  second job's reboot request.
* The second job stays pending forever, since it is waiting for the node
  to reboot, even though the node has already rebooted.

To avoid this race, we update the node's active features right before we
send the reboot request to the slurmd. That way, any future jobs
scheduled on that node with the feature will just wait for the node to
reboot and be ready *without* first sending a reboot request, since the
node already has the active feature.

Bug 12366