Allow --constraint with special operators without changeable features.
Slurm does not support using special operators "[]()|*" with changeable
features. However, when a node_features plugin was configured with
changeable features and a --constraint request was given with one of the
special operators but did not contain any changeable features, that job
request was incorrectly denied. This commit allows a job to request
non-changeable features using special operators while a node_features
plugin is configured.

Bug 10707