Revert "slurmrestd - fix job submission with auth/jwt."
This reverts commit ec14bcd1da7a.

Commit ec14bcd1da7a broke federated jobs by not sending the UID and GID
from the client to the origin slurmctld. If the UID and GID of the
job_desc message are not set, the origin slurmctld sets them to the UID
and GID of the credential.

However, the origin slurmctld passes the original message to the sibling
slurmctld's, and the original message did not have UID and GID set.
Because messages between siblings are over a persistent connection as
SlurmUser, at that point we no longer have the UID and GID of the job
submission, and the job submission request is rejected on the siblings.

By reverting commit ec14bcd1da7a we ensure that UID and GID are set in
the original job_desc message from the client.

Bug 12692