Log stepd to syslog if in foreground and SlurmdSyslogDebug is set
Match the stepd's syslog logging logic to the slurmd's in slurmd.c -->
_update_logging(). This also makes the stepd match the behavior
described in the docs for SlurmdSyslogDebug.

The issue was that the stepd logs would reach the syslog when the
slurmd was daemonized, but were unexpectedly being omitted from the
syslog when the slurmd was in the foreground with -D (not daemonzied).

Now that slurmd.service starts slurmd in the foreground with -D, this
was no longer an esoteric edge case.

Note that stepd logs will still not reach slurmd stdout/stderr if
the slurmd is in the foreground. This is known behavior.

See 5ef090561ff. The stepd was missed when SlurmdSyslogDebug was added.

Bug 10625