Remove SIGUSR1 handler from _agent().
This was used to interrupt blocked I/O while trying to shut the thread
down. However, nearly all I/O calls here are written to catch EINTR and
retry again without regard to the agent thread being shut down. And the
few spots that don't handle EINTR should be modified to do so.