Fix requeued cloud job getting BadConstraints
If the job was cancelled (e.g. PrologSlurmctld exit 1) before the nodes
were "powering up" (i.e. nodes are still "powered down") then the job and
nodes wouldn't be cleaned up correctly and wouldn't trigger a
batch_requeue_fini() which led to leaving information on the job and
causing the job to get BadConstraints.

Bug 12792