Remove duplicate printing of the jobid.
The jobid is already printed from libfile_bcast.la (_get_job_info).

It is actually printed 2 times in there once from the requested
(which could be hetjob, or an task from an array) and second from the
sbcast_cred found which is always the real job id whereas the
hetjob/array id could be different than the real id.

You will notice node_list is printed in _get_job_info() this is printed
for the same reason jobid is printed here as we could be comparing what
is in the sbcast_cred which will be printed right afterwards.