Handle timeout=NO_VAL16 (no timeout) in slurmscriptd
The default value for PrologEpilogTimeout is NO_VAL16 which is supposed
to be no timeout. For {Prolog|Epilog}Slurmctld which run through
slurmscriptd when passing NO_VAL16 (65534) as the timeout, this would
actually wait 65534 seconds instead of waiting indefinitely. This commit
handles NO_VAL16 and negative values as a timeout value in slurmscriptd.

Bug 13077