node_features/helpers - fix handling reboot with multiple job features
build_active_feature_bitmap2() is passed job features returned by
node_features_p_job_xlate(). build_active_feature_bitmap2() expects the
features to be comma-delimited. The helpers plugin did not
comma-delimit the features (but the knl_cray and knl_generic plugins do).
This meant that when a job requested multiple features with the helpers
plugin, all nodes would be sent a reboot request, even if they already
had the requred features.

This commit makes node_features_p_job_xlate() in the helpers plugin return
comma-delimited features, which in turn fixes the issue of rebooting
nodes even when they had required features.

Because this was an issue with the delimiter, this bug only happened
when the job requested multiple features.

Bug 12993