Doc - Only the slurmd daemon is affected by CoreSpecLimit and MemSpecLimit
In versions previous to slurm 21.08 slurmstepd was not effectively confined
in the system cgroup even if this was the intended behavior. In 21.08 and
newer this situation has been fixed and only slurmd in slurmd_cgroup.c adds
its pids to the system cgroup.

This situation is desierd because a limited slurmstepd could crash or
consume system resources when an mpi job has a large number of ranks.

We cannot mix the slurmd daemon and job related daemons into the same
cgroup.

Bug 12255