Fix logic around cleaning up track_script
Before this commit, the _track_script_rec_cleanup() thread would never be
signalled that the process was done cleaning up. That's because
track_script_thd_list is empty after track_script_flush() is called, but
track_script_killed() previously would look for the thread in
track_script_thd_list. This commits looks in the correct list to find and
signal the waiting thread.

Bug 12456