Check availability of cgroup controllers in slurmd
In jobacctgather we didn't check for the availability of the controllers in
slurmd. This lead slurmd to start just fine, causing spawned slurmstepds to
die when really initializing the plugin.

Just do a initialize in jobacctgather to check if the cgroup is properly
mounted. The init function just prepares the data structs and should not
affect anything.

slurmd already did a jobacctgather fini when shutting down.

Bug 8389