Clarify use of track_script_killed() in run_command()
At this point, status has not been set since we haven't called waitpid()
yet. So at this point, calling track_script_killed() just serves to break
from the while loop if the script was killed by track_script. So instead
of passing *status (which, from the point of view of run_command(), we
don't know if it was even set at this point), pass zero.

Bug 12456