Adapt pam_slurm_adopt to cgroup/v2
pam_slurm_adopt has been modified to work with both legacy and unified mode.
The logic when running under unified mode will work the same way than before
with one exception: when trying to randomly determine which job to pick for
the adoption, instead of reading the cgroup directories checking for the
one with earliest creation time, it will instead rely on the job id. So
accessing a node with more than one job id will adopt your process into the
lowest job id you own when the node is in unified mode.