Do not pthread_join detached threads
Currently, the only threads that use track_script are detached threads, so
track_script should not join those. We can also remove pthread_cancel,
since the reason for having pthread_cancel was to prevent pthread_join
from waiting forever.

In addition, track_script should not be in charge of joining threads. If
a thread needs to be joined, something outside of track_script should
handle exactly when that should be done.

Bug 12456