acct_gather_energy/xcc - close and destroy IPMI context after readings.
When a separate thread reads energy to respond to RPC request, a new context
is created. When the context isn't needed anymore, it needs to be closed and
destroyed. Othewise, memory and file descriptors could leak among other
issues.

In some circumstances, the number of leaked ipmi related fds can hit an
upper bound that results in slurmd receiving a SIGABRT.

Bug 12868