Testsuite - Fix reset_bash_prompt waiting for the right prompt
We don't know how many echoes will appear of the quoted version of the
prompt, so we need to wait for the final unquoted prompt.
We also need to keep that prompt into the expect buffer.
This is a continuation or fix of commits 9a5d21cc8b...2270c7458c.

Bug 12336