Make --mem=0 for a step not count against the job's memory allocation
Steps that request --mem=0 should be allocated all of the job's memory
on each node, but should not count against the job's memory allocation.
If they count against the job's allocation then it can allow steps that
request a memory amount greater than zero to overallocate the job's
memory.

Bug 13226