srun - Refactor --mpi=list option
It doesn't make sense to call MPI init just to print the available list
of MPI plugins. Clarify we don't really init the MPI plugin when we want
to print the list by moving the logic to it's own function in srun, plus
disabling and removing list value in plugin's init.

Bug 9395