Remove cgroup_allowed_devices_file.conf support
The cgroup's default policy in modern kernels and legacy mode (v1) is to
whitelist all devices by default. The contents of the file are not useful
anymore with the current code, since they are adding explicit allow rules
to a whitelist.

Also, it seems that having an administrator specifically set which devices
must be allowed/denied through Slurm is not convenient, unless they are
gres resources; in that case gres.conf can be used to deny specific devices
unless requested in the job.

Bug 5361