Cleanup nodes immediately for jobs cancelled before actually launching
Bug 11182 made it so that the Epilog wasn't run if the job never ran on
the nodes. Since the job never ran on the nodes, no epilog_completes
would clean up the job so job_epilog_complete() was called manually.
However, the nodes were cleaned up after EpilogSlurmcld finished -- ie.
the nodes stayed in a completing state until EpilogSlurmctld finished.
This actually goes against convention. When all epilog_completes finish
for a node, the node isn't completing anymore and is available to be
scheduled -- even though the job may still be completing waiting for
other epilogs and EpilogSlurmctld to complete.

This patch cleans up the nodes immediately at cancellation instead of
after the EpilogSlurmctld.

Bug 12792