Keep track of the slurm cgroup
We are already tracking all the other cgroups representing every node of
the hierarchy, namely root, uid, job, step and task cgroups. But the slurm
cgroup was not kept and when we needed it we had to create and load it again
each time. This will allow to get or set constrains from it directly.

Bug 12393