Docs - Update parameters that require restart vs reconfigure
Don't have scontrol docs list things that need a restart since
that list has changed and will probably continue to change.
Make sure the parameters that require a restart are properly
documented in the slurm.conf man page.
The parameters modified in this patch are not intended to be a
comprehensive list, but a starting point for further changes.

Bug 13237