Handle {Pro|Epi}logSlurmctld timing out
If {Pro|Epi}logSlurmctld runs longer than PrologEpilogTimeout, then the
script is killed.

Before this commit, if PrologSlurmctld timed out the job still ran. Now
if PrologSlurmctld times out, then an error is logged and the job will
not run. This is the same as if the PrologSlurmctld exits with a non-zero
exit status.

If EpilogSlurmctld times out, log an error.

Bug 11882