Make SPANK prolog/epilog calls respect PrologEpilogTimeout
This commit makes the SPANK plugin calls slurm_spank_job_prolog and
slurm_spank_job_epilog respect PrologEpilogTimeout.

Previously, it was unclear what the intended behavior of the code was.
It might have been to cap the timeout at 120 seconds but that did not
work. Since the documentation did not state that the timeout for these
SPANK plugin calls was capped, we do not try to "fix" that broken
behavior but instead make the behavior match the documentation.

Bug 11882