Wait for slurmstepd pid moved in cg2
We observed that sometimes the cgroup filesystem is slow, just as happens
in legacy mode. So we need to add the same check for slurmstepd to be moved
out of its cgroup before deleting it. In v1 we're actively reading a file
while here we opted for using inotify and poll.

Not waiting for this can cause the cleanup to not run properly and leave
stray cgroup directories around at random times.

Bug 12255