Steps may only overlap with steps that also used --overlap
In an allocation this allows for us to make things like the second step in

srun
followed by
srun --overlap

to pend instead of run as it has up to this date.
This was inconsistent with the following example:

srun --overlap
followed by
srun
will not run concurrently (expected).

Bug 12462