Manage the legacy cgroup levels in cgroup/v2
In cgroup/v1 we have a different hierarchy tree than in cgroup/v2, so there
are levels in the tree that are not existent in v2. Fix the functions where
cgroup_level_t is passed as a parameter converting it to the equivalent.

One example of such conversion needed is for commits:

- Inherit Correct limits for the UID cpuset cgroup
- Commit which adds CG_LEVEL_SLURM

Where trying to get the values for CG_LEVEL_SLURM would get an error.
Our CG_LEVEL_SLURM in cgroup/v2 is equivalent to CG_LEVEL_ROOT.

Bug 12255