Set affinity after setting up cgroups
In general, constraining any resource and after moving the pid of the task
around makes the constrains applied to stick with the cgroup, but not with
the task. This is also happening in the sched_setaffinity() case; if we
set the affinity and then move the pid to a different container then the
kernel will reset the affinity of the pid because the cpuset.cpus could have
changed.

This mainly affects cgroup/v2 since all the plugins deals with pids in the
same unified hierarchy. In this patch we change the order of affinity calls
vs cgroup directory setup.

Signed-off-by: Felip Moll <felip.moll@schedmd.com>

Bug 12255