Cleanup task_special when it is not used anymore in cgroup/v2
Proctrack works at step level, but in cgroup/v2 we are putting all the pids
in leaves, which means a task_special directory is created. If added to
proctrack we're using other plugins that move pids from task_special to
task_x directories, the first one can become empty. If this is the case a
restart of systemd can cause the directory to be vanished but a step will
still have it in its task_list.

To fix this we add logic to immediately cleanup the task_special directory
when it becomes empty.

Bug 12255