Revert "Do not add slurmstepd to the job cgroups"
This reverts commit 756207eaef3928d5a80127ab9093e4ea568d4b96.

We observed that not constraining the slurmstepd implies that spank plugins
will not either be constrained. Some customers use nvidia-smi in their
plugins so after this change all gpus were available when they expected not
to be. Moreover, having many unconstrained slurmstepds possibly from
different jobs can make one to affect the other, so it is desirable to only
affect the job if something bad happens rather than possibly affecting
other processes.

Bug 13189