Responsibility to move the stepd pid is plugin's not cgroup/v1
Unlike other plugins, the proctrack didn't move the stepd pid explictly into
its cgroup, but it was the cgroup/v1 plugin what was doing that implicit on
the step creation. Do like the others, and recognize if the stepd pid is the
one being inserted. The stepd pid must go into job cgroup and never into
step cgroup. We force here this logic to do the right thing.

Bug 13189