Allow slurm.conf nodes to be deleted from slurm.conf partitions
e.g.
NodeName=node[1-10] ...
PartitionName=p1 Nodes=node[1-10]

scontrol delete nodename=node1

Now that the partition remembers the original nodeline from the
slurm.conf, when the node is deleted and the partition's is rebuilt
(build_part_bitmap()), it will fail to find the deleted node and exit
before rebuilding the node. This changes it so that the partition will
rebuild and return an error if any nodes were not found and rebuild
part_ptr->nodes without the missing node. On startup, the slurmctld will
fatal if missing a node but on a dynamic removal, the partition will
ignore the missing node and remove the node from the original nodes
line.

Bug 13679