Prevent new elements from a job array from causing rerollups.
When a step record makes it to the dbd before the job start does
we checked the info filled in from the step which only included the
submit time where in an array it could cause the reroll to happen
needlessly in the past. Filling in the start time of the job as that
of the step fixes this problem.

Bug 13461

Co-authored-by: Scott Hilton <scott@schedmd.com>