Start using dbus to communicate with systemd
This requires some changes on how we initialize slurmstepd and slurmd.
From now on we are leaving slurmd in the cgroup where it was born. This
may have an impact when setting Core and MemorySpec* settings, if other
processes are on the same cgroup (e.g. user scope can have the bash
terminal or other user processes.). This shouldn't be a problem if slurmd
is started from a systemd unit and Delegate=yes set.

From now on, at startup the slurmd will try to initiate a transient systemd
unit, basically a scope. It will instruct systemd to initiate the scope for
us creating the cgroup directory and inserting a dummy pid into there, plus
setting Delegate=yes and TasksMax=infinity properties. The first one
instructs systemd to delegate the cgroup hierarchy associated to this scope
completely to us. TasksMax=infinity sets pids.max of the created cgroup to
'max', otherwise can be a predefined limit depending on system config, and
for now we do not manage the 'pids' controller from cgroup/v2, so it seems
reasonable to instruct systemd to manage it for us at the moment.

There's a failback mechanism which can be triggered if
IgnoreSystemdOnFailure is set in cgroup.conf, in that case slurmd will try
to create the scope manually, with mkdirs.

If one wants to completely ignore systemd and just do mkdirs, IgnoreSystemd
option from cgroup.conf is the option.

After the scope is created, if it ran successfuly with dbus, the dummy
process will be put into the new scope path ../system directory, and will
remain in order the cgroup to not be cleaned up ever by systemd. This could
change in the future if systemd decides to fix RemainAfterExit property for
services and add it for scopes. Actually even if Delegate=yes and this
property is set for a service, the cgroup will be erased by systemd when all
the pids are gone, when it is supposed not to. RemainAfterExit does not
exist for scopes, so that's why we have the dummy process.

After this setup is done, further slurmstepds will be moved into this
system/ directory on creation.

Note also that creating the scope through dbus has a delay, added to the
already known delay in the cgroup filesystem. We will do some retries while
checking if the cgroup directory of the scope is created.

Talking to dbus is done by an async communication, which means we cannot
expect a meaningful response from systemd/dbus. Even if we 'subscribe' for
JobStarted event in the dbus we won't get the certainity of the cgroup
directory being created.

Bug 12255