Testsuite - Improve 21.21 waiting for jobs before checking limits
In rare cases, later submitted jobs can run before earlier submitted
jobs (such as when an epilog is still in progress on the assigned
nodes), so we cannot rely on the last submitted job to be the one that
violates the limit.

Also, jobs are now pushed to a job_list variable so that they can be cleaned
up at the end of the test.

Bug 11692