Fix race while creating/removing cgroup directories in cgroup/v1
There was a missing xcgroup_lock in step creation which caused a race when
one step was setting up cgroups and another one removing them. This patch
adds the missing call to lock which will match with the xcgroup_lock already
present in the remove cgroup part of the code.

Bug 13423