Do not guess the node_name since we already have it in energy plugins
Calling to slurm_get_node_energy with NULL makes this api call to guess our
node name, but we already have it. Leave the code in the api as is just in
case somebody is still passing NULL but pass the node name from the energy
plugins, only locations where we're actually using this.

Bug 5518