Fix FAIL mail not being sent if a job was cancelled due to preemption
Fix regression beginning in 21.08.2 with commit 654d3af405 where
preempted jobs were no longer being sent FAIL mail. This makes sense
when the preempt mode is REQUEUE (since that is covered under the
REQUEUE mail type), but not when the preempt mode is CANCEL. So send
FAIL mail again when preempted if it's not a requeue (i.e., CANCEL).

The SUSPEND preempt mode does not apply here, since it doesn't appear
to trigger this code, and sending a mail every GANG schedule timeslice
swap makes no sense anyways.

Bug 13699