Add support for automounting controllers if IgnoreSystemd is set
init() will now call a new function depending on the IgnoreSystemd
parameter and will auto mount the supported controllers in the system
on slurmd directory. It will also setup the system.slice controllers.

This option is not really safe to use because systemd may eventually reset
the controllers. Normally this happens if no Delegate=yes unit exist in the
tree and an operation like reload or reset-failed is issued in systemd, but
we will keep it here for development purposes.

Bug 12255