Avoid bad config with CONFIG_MEMCG_SWAP=n and ConstrainSwapSpace=Yes
Fix that adding new cgroup plugin function cgroup_g_has_feature.

The memory controller has a features that can show up or not depending on
the system boot parameters or the kernel configuration (like others), but
specially CONFIG_MEMCG_SWAP is sometimes disabled at kernel level but still
the ConstrainSwapSpace is set in cgroup.conf, which causes issues.

Another way this problem may show up is on systems that need this in the
boot parameters: "cgroup_enable=memory swapaccount=1"

This function adds the possibility to query for a given feature to the
cgroup plugin, which will then query the system and return a boolean
depending on if this is available.

Bug 12211