Fix srun crash when het step fails to launch
When reserved ports are being used (ex. MpiParams=ports=...) srun can crash
when a het step fails to launch. For example (a and b don't exist):

$ sbatch -N1 --wrap="srun a : b"

The fix is to avoid further reserved port setup in _create_job_step() when
an error occurs in creating a step for a component job.

bug 13780