common/openapi - don't print method when freeing method-less entry list
This entry list is from the API spec and doesn't have any methods associated
with it, so don't try to associate one of the populated methods to it.

Also, because of a previously-fixed bug, args.method was always pointing to
a NULL object anyways, resulting in `INVLAID` as the method printed instead
of `UNKNOWN`.

Bug 13759