Remove double fork.
This was put in originally in commit 1ad826bf9f for burst buffer generic.

What this did is if the timeout was not set (so we wait forever)
we didn't wait at all since the now parent cpid would return immediately
leaving the it child (the script) as a child of init with no one ever
waiting on it.  So it basically never worked correctly, but up until
ba5c3323cc3 we never had anything that called run_command() without a
timeout (PrologEpilogTimeout).