Docs - clarify job_submit/lua environment example.
Move the example after the NOTE which clarifies that the
job_desc.environment is only available for batch jobs.
Improve the example content by showing how to differentiate batch vs
non-batch jobs inside a lua a script by checking the job_desc.script
field and improve the example format while here so that it is shown
inside a <pre> block.

Continuation of a5993ef2e1.

Bug 12992

Signed-off-by: Marcin Stolarek <cinek@schedmd.com>