mpi - mpi.conf configless mode
We only need to add mpi.conf to the list of files slurmd needs for
configless mode. This way, slurmd gets mpi.conf in the config cache
before trying to load it.

The rest of the logic remains the same, as slurmd is then able to load
MPI config without changes.

Bug 9395