Continuation of f084e13556
The node was destroyed passing the address of a variable instead of the
address of the list node, thus resulting on the origin list not having
its nodes' next links and list head/tail properly updated.

There are plenty of side effects on this, like being able to still
iterate over this list falling on effectively deleted nodes and crash,
or being able to access that node and free the data associated, that now
belongs to another list.

Bug 9395