Make proctrack_p_wait to really wait on cgroup emptiness
We were relying on the destruction of a cgroup to check for it being empty.
This is not ok, because if there are other cgroup plugins loaded we do not
want to try destroy the cgroup to not race with others. Also, multiple
signals sent to a step can cause wait to be called multiple times, so the
destroy must be avoided. Move it outside stepd_cleanup. This function is
only called once per step at termination.

Bug 13768

Signed-off-by: Felip Moll <felip.moll@schedmd.com>
Signed-off-by: Oriol Vilarrubi <jvilarru@schedmd.com>