Add ebpf library
This library will take care of creating the ebpf program which will be
injected into the cgroup in order to constrain devices.
Create also a macro to check for the linux bpf header and includes.
During configure if bpf.h is found then set WITH_BPF, which will be used
later in the cgroup plugin Makefile to decide whether to compile the v2
plugin or not.

Bug 12255