Avoid step destroy race adding cgroup locking mechanism in cgroup/v2
There was a race when multiple steps tried to remove the same job directory
or when a step creation and a termination coincided. We need to lock the
root directory before doing mkdir or rmdir of job_x dirs. This is related
to bug 13423 too but this one was only for v1.

Bug 12255