Move out of session scope when slurmd is started manually
If this is not done, then the slurmd will remain in the same cgroup than the
user session one, where for example we will have the terminal. If CoreSpec
or MemSpec limits are set, this will affect also the processes in this
cgroup, leaving your terminal with a limited number of resources.

This is undesirable, so in this case where slurmd is not started with
systemd, just move out our pid into the slurmstepd scope in a special
directory called slurmd. In that case nothing happens with systemd
because the scope is started manually with Delegate=yes, and we are not
expecting a systemd restart of this pid.

Bug 12255