Repurpose expand_nodesets() and rename it to nodespec_to_hostlist().
Repurpose the function so that:
- Previous two commits made it so that "ALL" keyword is managed in the
  function instead of the caller.
- Do not return a deranged string off of the hostlist, but instead
  the hostlist_t itself.

And adapt the current only one caller build_part_bitmap() to use it.

NOTE: Since the caller is expected to walk the hostlist, it's also been
necessary to convert hostlist_push_host() to hostlist_push() in two places:
the function itself and _add_nodes_with_feature(). Otherwise, when walking
with hostlist_shift() we could get ranges instead of a single host.

Bug 13099